recode TodoItem
